requirements.txt
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-api-logger==1.0.10